+++ services/app-web/src/pages/CMSDashboard/RateReviewsDashboard/index.ts
@@ -1,0 +1,1 @@
+export { RateReviewsDashboard } from './RateReviewsDashboard'
