+++ services/app-web/src/pages/CMSDashboard/SubmissionsDashboard/SubmissionsDashboard.tsx
@@ -1,0 +1,147 @@
+import React from 'react'
+import { packageName } from '../../../common-code/healthPlanFormDataType'
+import { base64ToDomain } from '../../../common-code/proto/healthPlanFormDataProto'
+import { SubmissionTypeRecord } from '../../../constants/healthPlanPackages'
+import { useAuth } from '../../../contexts/AuthContext'
+import { useIndexHealthPlanPackagesQuery } from '../../../gen/gqlClient'
+import { mostRecentDate } from '../../../common-code/dateHelpers'
+import styles from '../../StateDashboard/StateDashboard.module.scss'
+import { recordJSException } from '../../../otelHelpers/tracingHelper'
+import {
+    Loading,
+    HealthPlanPackageTable,
+    PackageInDashboardType,
+} from '../../../components'
+import { ErrorFailedRequestPage } from '../../Errors/ErrorFailedRequestPage'
+
+const SubmissionsDashboard = (): React.ReactElement => {
+    const { loggedInUser } = useAuth()
+    const { loading, data, error } = useIndexHealthPlanPackagesQuery({
+        fetchPolicy: 'network-only',
+    })
+
+    if (loading || !loggedInUser) {
+        return <Loading />
+    } else if (error) {
+        return (
+            <ErrorFailedRequestPage
+                error={error}
+                testID="submissions-dashboard"
+            />
+        )
+    }
+
+    const submissionRows: PackageInDashboardType[] = []
+    data?.indexHealthPlanPackages.edges
+        .map((edge) => edge.node)
+        .forEach((sub) => {
+            const currentRevision = sub.revisions[0]
+            const currentFormData = base64ToDomain(
+                currentRevision.node.formDataProto
+            )
+
+            // Errors - data handling
+            if (sub.status === 'DRAFT') {
+                recordJSException(
+                    `indexHealthPlanPackagesQuery: should not return draft submissions for CMS user. ID: ${sub.id}`
+                )
+                return
+            }
+            if (currentFormData instanceof Error) {
+                recordJSException(
+                    `indexHealthPlanPackagesQuery: Error decoding proto. ID: ${sub.id} Error message: ${currentFormData.message}`
+                )
+
+                return null
+            }
+
+            if (
+                currentRevision?.node?.submitInfo?.updatedAt === undefined &&
+                currentRevision?.node?.unlockInfo?.updatedAt === undefined
+            ) {
+                recordJSException(
+                    `indexHealthPlanPackagesQuery: Error finding submit and unlock dates for submissions in CMSDashboard. ID: ${sub.id}`
+                )
+            }
+
+            // Set package display data
+            let displayRateFormData = currentFormData
+            let lastUpdated = mostRecentDate([
+                currentRevision?.node?.submitInfo?.updatedAt,
+                currentRevision?.node?.unlockInfo?.updatedAt,
+            ])
+
+            if (sub.status === 'UNLOCKED') {
+                // Errors - data handling
+                const previousRevision = sub?.revisions[1]
+                const previousFormData = base64ToDomain(
+                    previousRevision?.node?.formDataProto
+                )
+                if (previousFormData instanceof Error) {
+                    recordJSException(
+                        `indexHealthPlanPackagesQuery: Error decoding proto for display of an unlocked submission. ID: ${sub.id} Error message: ${previousFormData.message}`
+                    )
+
+                    return
+                }
+
+                if (
+                    previousRevision?.node?.submitInfo?.updatedAt ===
+                        undefined &&
+                    previousRevision?.node?.unlockInfo?.updatedAt === undefined
+                ) {
+                    recordJSException(
+                        `indexHealthPlanPackagesQuery: Error finding submit and unlock dates of an unlocked submission. ID: ${sub.id}`
+                    )
+                }
+
+                // reset package display data since unlock submissions rely on previous revision data
+                displayRateFormData = previousFormData
+                lastUpdated = mostRecentDate([
+                    currentRevision?.node?.submitInfo?.updatedAt,
+                    currentRevision?.node?.unlockInfo?.updatedAt,
+                    previousRevision?.node?.submitInfo?.updatedAt,
+                    previousRevision?.node?.unlockInfo?.updatedAt,
+                ])
+            }
+
+            if (!lastUpdated) {
+                recordJSException(
+                    `CMSDashboard: Cannot find valid last updated date from submit and unlock info. Falling back to current revision's last edit timestamp. ID: ${sub.id}`
+                )
+                lastUpdated = new Date(currentFormData.updatedAt)
+            }
+            const programs = sub.state.programs
+
+            submissionRows.push({
+                id: sub.id,
+                name: packageName(
+                    displayRateFormData.stateCode,
+                    displayRateFormData.stateNumber,
+                    displayRateFormData.programIDs,
+                    programs
+                ),
+                programs: programs.filter((program) =>
+                    displayRateFormData.programIDs.includes(program.id)
+                ),
+                submittedAt: sub.initiallySubmittedAt,
+                status: sub.status,
+                updatedAt: lastUpdated,
+                submissionType:
+                    SubmissionTypeRecord[displayRateFormData.submissionType],
+                stateName: sub.state.name,
+            })
+        })
+
+    return (
+        <section className={styles.panel}>
+            <HealthPlanPackageTable
+                tableData={submissionRows}
+                user={loggedInUser}
+                showFilters
+            />
+        </section>
+    )
+}
+
+export { SubmissionsDashboard }
