+++ services/app-web/src/pages/CMSDashboard/SubmissionsDashboard/index.ts
@@ -1,0 +1,1 @@
+export { SubmissionsDashboard } from './SubmissionsDashboard'
