+++ services/app-web/src/pages/CMSDashboard/index.ts
@@ -1,0 +1,3 @@
+export { RateReviewsDashboard } from './RateReviewsDashboard'
+export { SubmissionsDashboard } from './SubmissionsDashboard'
+export { CMSDashboard } from './CMSDashboard'
